64b initial pc, load icnt from checkpoint for mtime

+++ top.hh
@@ -280,7 +280,7 @@ static inline uint8_t *mmap4G() {
 
 static inline
 void reset_core(std::unique_ptr<Vcore_l1d_l1i> &tb, uint64_t &cycle,
-		uint32_t init_pc) {
+		uint64_t init_pc) {
   for(; (cycle < 4) && !Verilated::gotFinish(); ++cycle) {
     tb->mem_rsp_valid = 0;
     tb->monitor_ack = 0;
